src/main/java/pl/horus/Wall.java
package pl.horus;

import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

public class Wall implements Structure, CompositeBlock {

    private List<Block> blocks;

    public Wall(List<Block> blocks) {
        this.blocks = blocks;
    }

    @Override
    public Optional<Block> findBlockByColor(String color) {
        return blocks.stream()
                .filter(block -> block.getColor().equals(color))
                .findFirst();
    }

    @Override
    public List<Block> findBlocksByMaterial(String material) {
        return blocks.stream()
                .filter(block -> block.getMaterial().equals(material))
                .collect(Collectors.toList());
    }

    @Override
    public int count() {
        return blocks.size();
    }

    @Override
    public List<Block> getBlocks() {
        return blocks;
    }

    @Override
    public String getColor() {
        if (!blocks.isEmpty()) {
            return blocks.get(0).getColor();
        } else {
            return "No color";
        }
    }

    @Override
    public String getMaterial() {
        if (!blocks.isEmpty()) {
            return blocks.get(0).getMaterial();
        } else {
            return "Lack of material";
        }
    }
}